ci: updates to CI

+++ src/main/resources/db/migration/V0.0.1__Initial.sql
@@ -1,7 +1,7 @@
 -- Table Definition ----------------------------------------------
 
 CREATE TABLE "w_cars" (
-    "car_id" numeric PRIMARY KEY,
+    "car_id" serial PRIMARY KEY,
     "name" text NOT NULL,
     "description" text NOT NULL
 );
